+++ CodeWars/Fundamentals/BallUpwards/Java/src/Ball.java
@@ -1,0 +1,10 @@
+/**
+ * Created by cajetan on 24/07/2017.
+ */
+
+public class Ball {
+
+    public static int maxBall(int v0) {
+        return (int)Math.round(v0/3.6/0.981);
+    }
+}